Fix Float/Double type mismatches in AudioVisualizerView
Convert PI constant to Float and explicitly cast sin() results to Float
to resolve compilation errors where Double values were being used in
Float contexts. Fixes type mismatches in pulse value calculations for
both glow halos and central orb rendering.

+++ app/app/src/main/java/uk/co/mrsheep/halive/ui/AudioVisualizerView.kt
@@ -362,7 +362,7 @@ class AudioVisualizerView @JvmOverloads constructor(
         pulseFrequency: Float,
         opacity: Float
     ) {
-        val pulseValue = (sin(elapsedSeconds * 2f * PI * pulseFrequency) + 1f) / 2f
+        val pulseValue = ((sin(elapsedSeconds * 2f * PI.toFloat() * pulseFrequency) + 1f) / 2f).toFloat()
         val haloPaints = arrayOf(paintHalo1, paintHalo2, paintHalo3)
 
         for (i in haloPaints.indices) {
@@ -424,7 +424,7 @@ class AudioVisualizerView @JvmOverloads constructor(
         opacity: Float,
         colorBlend: Float
     ) {
-        val pulseValue = (sin(elapsedSeconds * 2f * PI * pulseFrequency) + 1f) / 2f
+        val pulseValue = ((sin(elapsedSeconds * 2f * PI.toFloat() * pulseFrequency) + 1f) / 2f).toFloat()
         val orbRadius = 15f + (pulseValue * 10f)
 
         // Draw background halo
